+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 - Adds Github Actions for testing and pre-commit hooks.
+- Adds more unit tests for thirdpary email password recipe
+- Adds test to jwt recipe
+- Adds test to opendID recipe
 
 ## [0.5.7] - 2022-04-23
 - Adds functions to delete passwordless user info in recipes that have passwordless users.